requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
groq
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
qrcode==8.0
gunicorn==23.0.0
toolhouse
# twilio